1. 治疗图片接口调整到治疗记录模块 2. 增加治疗前，治疗后多张图片上传接口 3. 治疗记录上传成功后返回记录ID

+++ patient-manage/src/main/java/com/ruoyi/patient/service/ITreatmentRecordService.java
@@ -33,7 +33,7 @@ public interface ITreatmentRecordService
      * @param treatmentRecord 治疗记录
      * @return 结果
      */
-    public int insertTreatmentRecord(TreatmentRecord treatmentRecord);
+    public long insertTreatmentRecord(TreatmentRecord treatmentRecord);
 
     /**
      * 修改治疗记录
